+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\files\getDirectory
 - Adds oihana\files\getFileExtension
 - Adds oihana\files\getHomeDirectory
+- Adds oihana\files\getRoot
 - Adds oihana\files\getTemporaryDirectory
 - Adds oihana\files\getTimestampedDirectory
 - Adds oihana\files\getTimestampedFile
@@ -44,9 +45,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Adds oihana\files\path\canonicalizePath
 - Adds oihana\files\path\extractCanonicalParts
-- Adds oihana\files\path\isAbsolute
+- Adds oihana\files\path\isAbsolutePath
 - Adds oihana\files\path\isBasePath
 - Adds oihana\files\path\isLocalPath
+- Adds oihana\files\path\isRelativePath
 - Adds oihana\files\path\joinPaths
 - Adds oihana\files\path\normalizePath
 - Adds oihana\files\path\splitPath